refactor server src files into directories
