+++ server/index.js
@@ -2,7 +2,7 @@ const express = require("express");
 const mongoose = require("mongoose");
 const cookieParser = require("cookie-parser");
 const cors = require("cors");
-const userRoutes = require("./src/routes/userRoutes");
+const userRoutes = require("./src/user/routes");
 
 mongoose
     .connect(
