--- server/src/middlewares/authenticateDeveloper.js
@@ -1,21 +1,0 @@
-const jwt = require("jsonwebtoken");
-
-const authenticate = (req, res, next) => {
-    const authHeader = req.headers.authorization;
-    const token = authHeader && authHeader.split(" ")[1];
-
-    if (!token) {
-        return res.status(401).json({ message: "Unauthorized" });
-    }
-
-    try {
-        const decoded = jwt.verify(token, "developer-secret-key");
-        req.developer = decoded.developer;
-
-        next();
-    } catch (error) {
-        return res.status(401).json({ message: "Invalid token" });
-    }
-};
-
-module.exports = authenticate;
