+++ server/src/user/controllers.js
@@ -1,6 +1,6 @@
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
-const User = require("../models/userModel");
+const User = require("./model");
 
 async function register(req, res) {
     try {
